moved social media share components
